Move and rename output worker base class
